RefactorTemplateSettingsPage tests
