+++ site/src/pages/TemplateSettingsPage/TemplateGeneralSettingsPage/TemplateSettingsPageView.stories.tsx
@@ -1,40 +1,33 @@
-import { action } from "@storybook/addon-actions";
-import { Story } from "@storybook/react";
 import { mockApiError, MockTemplate } from "testHelpers/entities";
-import {
-  TemplateSettingsPageView,
-  TemplateSettingsPageViewProps,
-} from "./TemplateSettingsPageView";
+import { TemplateSettingsPageView } from "./TemplateSettingsPageView";
+import type { Meta, StoryObj } from "@storybook/react";
 
-export default {
+const meta: Meta<typeof TemplateSettingsPageView> = {
   title: "pages/TemplateSettingsPageView",
   component: TemplateSettingsPageView,
   args: {
     template: MockTemplate,
-    onSubmit: action("onSubmit"),
-    onCancel: action("cancel"),
   },
 };
 
-const Template: Story<TemplateSettingsPageViewProps> = (args) => (
-  <TemplateSettingsPageView {...args} />
-);
+export default meta;
+type Story = StoryObj<typeof TemplateSettingsPageView>;
 
-export const Example = Template.bind({});
-Example.args = {};
+export const Example: Story = {};
 
-export const SaveTemplateSettingsError = Template.bind({});
-SaveTemplateSettingsError.args = {
-  submitError: mockApiError({
-    message: 'Template "test" already exists.',
-    validations: [
-      {
-        field: "name",
-        detail: "This value is already in use and should be unique.",
-      },
-    ],
-  }),
-  initialTouched: {
-    allow_user_cancel_workspace_jobs: true,
+export const SaveTemplateSettingsError: Story = {
+  args: {
+    submitError: mockApiError({
+      message: 'Template "test" already exists.',
+      validations: [
+        {
+          field: "name",
+          detail: "This value is already in use and should be unique.",
+        },
+      ],
+    }),
+    initialTouched: {
+      allow_user_cancel_workspace_jobs: true,
+    },
   },
 };
